Data.Finite.Integral

Description

Documentation

class Integral a => SaneIntegral a

A class of datatypes that faithfully represent a sub-range of Integer that includes 0. A valid instance must obey the following laws:

fromInteger must be a retract of toInteger:

fromInteger (toInteger a) == a

Restricted to the range [0, Limit] (with Nothing understood as positive infinity), fromInteger must be an inverse of toInteger:

limited i ==> toInteger (fromInteger i) == i

where:

limited i = case limit of
    Just l -> 0 <= i && i <= l
    Nothing -> 0 <= i

WARNING: violating the above constraint in particular breaks type safety.

The implementations of Ord, Enum, Num, Integral must be compatible with that of Integer, whenever all arguments and results fall within [0, Limit], for example:

limited i && limited j && limited k && (i * j == k) ==> (fromInteger i * fromInteger j == fromInteger k)

Methods modAdd, modSub, and modMul implement modular addition, multiplication, and subtraction. The default implementation is via Integer, but a faster implementation can be provided instead. If provided, the implementation must be correct for moduli in range [1, Limit].

WARNING: a naive implementaton is prone to arithmetic overflow and may produce invalid results for moduli close to Limit.

type Limited a (n :: Nat) = LeqMaybe n (Limit a)

Ensures that the value of n is representable in type a (which should be a SaneIntegral).

class KnownIntegral a (n :: Nat)

This class asserts that the value of n is known at runtime, and that it is representable in type a (which should be a SaneIntegral).

At runtime it acts like an implicit parameter of type a, much like KnownNat is an implicit parameter of type Integer.

Minimal complete definition

intVal_

intVal :: forall (n :: Nat) a proxy. KnownIntegral a n => proxy n -> a

Reflect a type-level number into a term.

withIntegral :: forall a (n :: Nat) r proxy1 proxy2. (SaneIntegral a, KnownIntegral a n) => proxy1 a -> proxy2 n -> (KnownNat n => r) -> r

Recover a KnownNat constraint from a KnownIntegral constraint.

data Finite a (n :: Nat)

Finite number type. The type Finite a n is inhabited by exactly n values from type a, in the range [0, n) including 0 but excluding n. a must be an instance of SaneIntegral to use this type. Invariants:

getFinite x < intVal x

getFinite x >= 0

packFinite :: forall (n :: Nat) a. (SaneIntegral a, KnownIntegral a n) => a -> Maybe (Finite a n)

Convert an a into a Finite a, returning Nothing if the input is out of bounds.

packFiniteProxy :: forall (n :: Nat) a proxy. (SaneIntegral a, KnownIntegral a n) => proxy n -> a -> Maybe (Finite a n)

Same as packFinite but with a proxy argument to avoid type signatures.

finite :: forall (n :: Nat) a. (SaneIntegral a, KnownIntegral a n) => a -> Finite a n

Convert an a into a Finite a, throwing an error if the input is out of bounds.

finiteProxy :: forall (n :: Nat) a proxy. (SaneIntegral a, KnownIntegral a n) => proxy n -> a -> Finite a n

Same as finite but with a proxy argument to avoid type signatures.

getFinite :: forall (n :: Nat) a. Finite a n -> a

Convert a Finite a into the corresponding a.

finites :: forall (n :: Nat) a. (SaneIntegral a, KnownIntegral a n) => [Finite a n]

Generate an ascending list of length n of all elements of Finite a n.

finitesProxy :: forall (n :: Nat) a proxy. (SaneIntegral a, KnownIntegral a n) => proxy n -> [Finite a n]

Same as finites but with a proxy argument to avoid type signatures.

modulo :: forall (n :: Nat) a. (SaneIntegral a, KnownIntegral a n) => a -> Finite a n

Produce the Finite that is congruent to the given integer modulo n.

moduloProxy :: forall (n :: Nat) a proxy. (SaneIntegral a, KnownIntegral a n) => proxy n -> a -> Finite a n

Same as modulo but with a proxy argument to avoid type signatures.

equals :: forall (n :: Nat) (m :: Nat) a. Eq a => Finite a n -> Finite a m -> Bool infix 4

Test two different types of finite numbers for equality.

cmp :: forall (n :: Nat) (m :: Nat) a. Ord a => Finite a n -> Finite a m -> Ordering

Compare two different types of finite numbers.

natToFinite :: forall (n :: Nat) (m :: Nat) a proxy. (SaneIntegral a, KnownIntegral a n, Limited a m, (n + 1) <= m) => proxy n -> Finite a m

Convert a type-level literal into a Finite.

weaken :: forall (n :: Natural) a. Limited a (n + 1) => Finite a n -> Finite a (n + 1)

Add one inhabitant in the end.

strengthen :: forall (n :: Nat) a. (SaneIntegral a, KnownIntegral a n) => Finite a (n + 1) -> Maybe (Finite a n)

Remove one inhabitant from the end. Returns Nothing if the input was the removed inhabitant.

shift :: forall (n :: Natural) a. (SaneIntegral a, Limited a (n + 1)) => Finite a n -> Finite a (n + 1)

Add one inhabitant in the beginning, shifting everything up by one.

unshift :: forall (n :: Natural) a. SaneIntegral a => Finite a (n + 1) -> Maybe (Finite a n)

Remove one inhabitant from the beginning, shifting everything down by one. Returns Nothing if the input was the removed inhabitant.

weakenN :: forall (n :: Nat) (m :: Nat) a. (n <= m, Limited a m) => Finite a n -> Finite a m

Add multiple inhabitants in the end.

strengthenN :: forall (n :: Nat) (m :: Nat) a. (SaneIntegral a, KnownIntegral a m, Limited a m) => Finite a n -> Maybe (Finite a m)

Remove multiple inhabitants from the end. Returns Nothing if the input was one of the removed inhabitants.

shiftN :: forall (n :: Nat) (m :: Nat) a. (SaneIntegral a, KnownIntegral a n, KnownIntegral a m, n <= m) => Finite a n -> Finite a m

Add multiple inhabitants in the beginning, shifting everything up by the amount of inhabitants added.

unshiftN :: forall (n :: Nat) (m :: Nat) a. (SaneIntegral a, KnownIntegral a n, KnownIntegral a m, Limited a m) => Finite a n -> Maybe (Finite a m)

Remove multiple inhabitants from the beginning, shifting everything down by the amount of inhabitants removed. Returns Nothing if the input was one of the removed inhabitants.

weakenProxy :: forall (n :: Natural) (k :: Natural) a proxy. Limited a (n + k) => proxy k -> Finite a n -> Finite a (n + k)

strengthenProxy :: forall (n :: Nat) (k :: Natural) a proxy. (SaneIntegral a, KnownIntegral a n) => proxy k -> Finite a (n + k) -> Maybe (Finite a n)

shiftProxy :: forall (n :: Natural) (k :: Nat) a proxy. (SaneIntegral a, KnownIntegral a k, Limited a (n + k)) => proxy k -> Finite a n -> Finite a (n + k)

unshiftProxy :: forall (n :: Natural) (k :: Nat) a proxy. (SaneIntegral a, KnownIntegral a k) => proxy k -> Finite a (n + k) -> Maybe (Finite a n)

add :: forall (n :: Natural) (m :: Natural) a. (SaneIntegral a, Limited a (n + m)) => Finite a n -> Finite a m -> Finite a (n + m)

Add two Finites.

sub :: forall (n :: Nat) (m :: Nat) a. SaneIntegral a => Finite a n -> Finite a m -> Either (Finite a m) (Finite a n)

Subtract two Finites. Returns Left for negative results, and Right for positive results. Note that this function never returns Left 0.

multiply :: forall (n :: Natural) (m :: Natural) a. (SaneIntegral a, Limited a (n * m)) => Finite a n -> Finite a m -> Finite a (n * m)

Multiply two Finites.

combineSum :: forall (n :: Nat) (m :: Natural) a. (SaneIntegral a, KnownIntegral a n, Limited a (n + m)) => Either (Finite a n) (Finite a m) -> Finite a (n + m)

Left-biased (left values come first) disjoint union of finite sets.

combineZero :: Void -> Finite a 0

Witness that combineSum preserves units: 0 is the unit of +, and Void is the unit of Either.

combineProduct :: forall (n :: Nat) (m :: Natural) a. (SaneIntegral a, KnownIntegral a n, Limited a (n * m)) => (Finite a n, Finite a m) -> Finite a (n * m)

fst-biased (fst is the inner, and snd is the outer iteratee) product of finite sets.

combineOne :: (SaneIntegral a, Limited a 1) => () -> Finite a 1

Witness that combineProduct preserves units: 1 is the unit of *, and () is the unit of (,).

combineExponential :: forall (n :: Nat) (m :: Nat) a. (SaneIntegral a, KnownIntegral a m, KnownIntegral a n, Limited a (m ^ n)) => (Finite a n -> Finite a m) -> Finite a (m ^ n)

Product of n copies of a finite set of size m, biased towards the lower values of the argument (colex order).

separateSum :: forall (n :: Nat) (m :: Natural) a. (SaneIntegral a, KnownIntegral a n) => Finite a (n + m) -> Either (Finite a n) (Finite a m)

Take a Left-biased disjoint union apart.

separateZero :: SaneIntegral a => Finite a 0 -> Void

Witness that separateSum preserves units: 0 is the unit of +, and Void is the unit of Either.

Also witness that a Finite a 0 is uninhabited.

separateProduct :: forall (n :: Nat) (m :: Natural) a. (SaneIntegral a, KnownIntegral a n) => Finite a (n * m) -> (Finite a n, Finite a m)

Take a fst-biased product apart.

separateOne :: Finite a 1 -> ()

separateExponential :: forall (n :: Natural) (m :: Nat) a. (SaneIntegral a, KnownIntegral a m) => Finite a (m ^ n) -> Finite a n -> Finite a m

Take a product of n copies of a finite set of size m apart, biased towards the lower values of the argument (colex order).

castFinite :: forall b a (n :: Nat). (SaneIntegral a, SaneIntegral b, Limited b n) => Finite a n -> Finite b n

Convert a Finite between different SaneIntegral numeric types.

isValidFinite :: forall (n :: Nat) a. (Ord a, Num a, KnownIntegral a n) => Finite a n -> Bool

Verifies that a given Finite is valid. Should always return True unless you bring the Data.Finite.Internal.Finite constructor into the scope, or use unsafeCoerce or other nasty hacks.