Data.Finite.Internal

Description

Documentation

type Finite = Finite Integer

Finite number type. The type Finite n is inhabited by exactly n values in the range [0, n) including 0 but excluding n. Invariants:

getFinite x < natVal x

getFinite x >= 0

pattern Finite :: Integer -> Finite n

finite :: forall (n :: Nat). KnownNat n => Integer -> Finite n

Convert an Integer into a Finite, throwing an error if the input is out of bounds.

getFinite :: forall (n :: Nat). Finite n -> Integer

Convert a Finite into the corresponding Integer.