Data.Finite.Internal.Integral

Description

Documentation

class Integral a => SaneIntegral a where

A class of datatypes that faithfully represent a sub-range of Integer that includes 0. A valid instance must obey the following laws:

fromInteger must be a retract of toInteger:

fromInteger (toInteger a) == a

Restricted to the range [0, Limit] (with Nothing understood as positive infinity), fromInteger must be an inverse of toInteger:

limited i ==> toInteger (fromInteger i) == i

where:

limited i = case limit of
    Just l -> 0 <= i && i <= l
    Nothing -> 0 <= i

WARNING: violating the above constraint in particular breaks type safety.

The implementations of Ord, Enum, Num, Integral must be compatible with that of Integer, whenever all arguments and results fall within [0, Limit], for example:

limited i && limited j && limited k && (i * j == k) ==> (fromInteger i * fromInteger j == fromInteger k)

Methods modAdd, modSub, and modMul implement modular addition, multiplication, and subtraction. The default implementation is via Integer, but a faster implementation can be provided instead. If provided, the implementation must be correct for moduli in range [1, Limit].

WARNING: a naive implementaton is prone to arithmetic overflow and may produce invalid results for moduli close to Limit.

Minimal complete definition

Nothing

Associated Types

type Limit a :: Maybe Nat

Methods

modAdd :: a -> a -> a -> a

Given n > 0, 0 <= a < n, and 0 <= b < n, modAdd n a b computes (a + b) ` mod ` n.

modSub :: a -> a -> a -> a

Given n > 0, 0 <= a < n, and 0 <= b < n, modSub n a b computes (a - b) ` mod ` n.

modMul :: a -> a -> a -> a

Given n > 0, 0 <= a < n, and 0 <= b < n, modMul n a b computes (a * b) ` mod ` n.

unsafeWithLimited :: forall proxy1 proxy2 (n :: Nat) r. proxy1 a -> proxy2 n -> (Limited a n => r) -> r

Unsafely obtain evidence that n <= Limit a. When Limit is Nothing there is no evidence to obtain, and \_ _ k -> k is a valid implementation. When Limit is a Just, the default implementation should work.

default unsafeWithLimited :: forall (n :: Nat) r (lim :: Nat) proxy1 proxy2. Limit a ~ 'Just lim => proxy1 a -> proxy2 n -> (Limited a n => r) -> r

type Limited a (n :: Nat) = LeqMaybe n (Limit a)

Ensures that the value of n is representable in type a (which should be a SaneIntegral).

class KnownIntegral a (n :: Nat)

This class asserts that the value of n is known at runtime, and that it is representable in type a (which should be a SaneIntegral).

At runtime it acts like an implicit parameter of type a, much like KnownNat is an implicit parameter of type Integer.

Minimal complete definition

intVal_

intVal :: forall (n :: Nat) a proxy. KnownIntegral a n => proxy n -> a

Reflect a type-level number into a term.

withIntegral :: forall a (n :: Nat) r proxy1 proxy2. (SaneIntegral a, KnownIntegral a n) => proxy1 a -> proxy2 n -> (KnownNat n => r) -> r

Recover a KnownNat constraint from a KnownIntegral constraint.

withLimited :: forall a (n :: Nat) r proxy1 proxy2. (SaneIntegral a, KnownIntegral a n) => proxy1 a -> proxy2 n -> (Limited a n => r) -> r

Recover a Limited constraint from a KnownIntegral constraint.

newtype Finite a (n :: Nat)

Finite number type. The type Finite a n is inhabited by exactly n values from type a, in the range [0, n) including 0 but excluding n. a must be an instance of SaneIntegral to use this type. Invariants:

getFinite x < intVal x

getFinite x >= 0

Constructors

Finite a

finite :: forall (n :: Nat) a. (SaneIntegral a, KnownIntegral a n) => a -> Finite a n

Convert an a into a Finite a, throwing an error if the input is out of bounds.

getFinite :: forall (n :: Nat) a. Finite a n -> a

Convert a Finite a into the corresponding a.