Data.Finite

Description

Documentation

type Finite = Finite Integer

Finite number type. The type Finite n is inhabited by exactly n values in the range [0, n) including 0 but excluding n. Invariants:

getFinite x < natVal x

getFinite x >= 0

packFinite :: forall (n :: Nat). KnownNat n => Integer -> Maybe (Finite n)

Convert an Integer into a Finite, returning Nothing if the input is out of bounds.

packFiniteProxy :: forall (n :: Nat) proxy. KnownNat n => proxy n -> Integer -> Maybe (Finite n)

Same as packFinite but with a proxy argument to avoid type signatures.

finite :: forall (n :: Nat). KnownNat n => Integer -> Finite n

Convert an Integer into a Finite, throwing an error if the input is out of bounds.

finiteProxy :: forall (n :: Nat) proxy. KnownNat n => proxy n -> Integer -> Finite n

Same as finite but with a proxy argument to avoid type signatures.

getFinite :: forall (n :: Nat). Finite n -> Integer

Convert a Finite into the corresponding Integer.

finites :: forall (n :: Nat). KnownNat n => [Finite n]

Generate a list of length n of all elements of Finite n.

finitesProxy :: forall (n :: Nat) proxy. KnownNat n => proxy n -> [Finite n]

Same as finites but with a proxy argument to avoid type signatures.

modulo :: forall (n :: Nat). KnownNat n => Integer -> Finite n

Produce the Finite that is congruent to the given integer modulo n.

moduloProxy :: forall (n :: Nat) proxy. KnownNat n => proxy n -> Integer -> Finite n

Same as modulo but with a proxy argument to avoid type signatures.

equals :: forall (n :: Nat) (m :: Nat). Finite n -> Finite m -> Bool infix 4

Test two different types of finite numbers for equality.

cmp :: forall (n :: Nat) (m :: Nat). Finite n -> Finite m -> Ordering

Compare two different types of finite numbers.

natToFinite :: forall (n :: Nat) (m :: Nat) proxy. (KnownNat n, KnownNat m, (n + 1) <= m) => proxy n -> Finite m

Convert a type-level literal into a Finite.

weaken :: forall (n :: Nat). Finite n -> Finite (n + 1)

Add one inhabitant in the end.

strengthen :: forall (n :: Nat). KnownNat n => Finite (n + 1) -> Maybe (Finite n)

Remove one inhabitant from the end. Returns Nothing if the input was the removed inhabitant.

shift :: forall (n :: Nat). Finite n -> Finite (n + 1)

Add one inhabitant in the beginning, shifting everything up by one.

unshift :: forall (n :: Natural). Finite (n + 1) -> Maybe (Finite n)

Remove one inhabitant from the beginning, shifting everything down by one. Returns Nothing if the input was the removed inhabitant.

weakenN :: forall (n :: Nat) (m :: Nat). n <= m => Finite n -> Finite m

Add multiple inhabitants in the end.

strengthenN :: forall (n :: Nat) (m :: Nat). KnownNat m => Finite n -> Maybe (Finite m)

Remove multiple inhabitants from the end. Returns Nothing if the input was one of the removed inhabitants.

shiftN :: forall (n :: Nat) (m :: Nat). (KnownNat n, KnownNat m, n <= m) => Finite n -> Finite m

Add multiple inhabitants in the beginning, shifting everything up by the amount of inhabitants added.

unshiftN :: forall (n :: Nat) (m :: Nat). (KnownNat n, KnownNat m) => Finite n -> Maybe (Finite m)

Remove multiple inhabitants from the beginning, shifting everything down by the amount of inhabitants removed. Returns Nothing if the input was one of the removed inhabitants.

weakenProxy :: forall (n :: Nat) (k :: Natural) proxy. proxy k -> Finite n -> Finite (n + k)

strengthenProxy :: forall (n :: Nat) (k :: Natural) proxy. KnownNat n => proxy k -> Finite (n + k) -> Maybe (Finite n)

shiftProxy :: forall (n :: Nat) (k :: Nat) proxy. KnownNat k => proxy k -> Finite n -> Finite (n + k)

unshiftProxy :: forall (n :: Natural) (k :: Nat) proxy. KnownNat k => proxy k -> Finite (n + k) -> Maybe (Finite n)

add :: forall (n :: Nat) (m :: Nat). Finite n -> Finite m -> Finite (n + m)

Add two Finites.

sub :: forall (n :: Nat) (m :: Nat). Finite n -> Finite m -> Either (Finite m) (Finite n)

Subtract two Finites. Returns Left for negative results, and Right for positive results. Note that this function never returns Left 0.

multiply :: forall (n :: Nat) (m :: Nat). Finite n -> Finite m -> Finite (n * m)

Multiply two Finites.

combineSum :: forall (n :: Nat) (m :: Nat). KnownNat n => Either (Finite n) (Finite m) -> Finite (n + m)

Left-biased (left values come first) disjoint union of finite sets.

combineZero :: Void -> Finite 0

Witness that combineSum preserves units: 0 is the unit of +, and Void is the unit of Either.

combineProduct :: forall (n :: Nat) (m :: Nat). KnownNat n => (Finite n, Finite m) -> Finite (n * m)

fst-biased (fst is the inner, and snd is the outer iteratee) product of finite sets.

combineOne :: () -> Finite 1

Witness that combineProduct preserves units: 1 is the unit of *, and () is the unit of (,).

combineExponential :: forall (n :: Nat) (m :: Nat). (KnownNat m, KnownNat n) => (Finite n -> Finite m) -> Finite (m ^ n)

Product of n copies of a finite set of size m, biased towards the lower values of the argument (colex order).

separateSum :: forall (n :: Nat) (m :: Natural). KnownNat n => Finite (n + m) -> Either (Finite n) (Finite m)

Take a Left-biased disjoint union apart.

separateZero :: Finite 0 -> Void

Witness that separateSum preserves units: 0 is the unit of +, and Void is the unit of Either.

Also witness that a Finite 0 is uninhabited.

separateProduct :: forall (n :: Nat) (m :: Natural). KnownNat n => Finite (n * m) -> (Finite n, Finite m)

Take a fst-biased product apart.

separateOne :: Finite 1 -> ()

separateExponential :: forall (n :: Natural) (m :: Nat). KnownNat m => Finite (m ^ n) -> Finite n -> Finite m

Take a product of n copies of a finite set of size m apart, biased towards the lower values of the argument (colex order).

isValidFinite :: forall (n :: Nat). KnownNat n => Finite n -> Bool

Verifies that a given Finite is valid. Should always return True unless you bring the Data.Finite.Internal.Finite constructor into the scope, or use unsafeCoerce or other nasty hacks.